# openapi.yaml
openapi: 3.0.0
info:
  title: Social Network API
  description: Serwis backendowy z bazami PostgreSQL oraz MongoDB do zarządzania feedem, postami i interakcjami (sieć społecznościowa).
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Główny API Gateway
paths:
  /posts:
    post:
      summary: Utworzenie postu i rozgłoszenie do obserwujących
      description: Tworzy rekord zapisu postu w PostgreSQL oraz zdenormalizowane entry w systemie MongoDB dla całego feedu obserwatorów.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - authorId
                - bodyPreview
              properties:
                authorId:
                  type: integer
                  example: 1
                bodyPreview:
                  type: string
                  example: "Krótka zajawka postu"
                attachments:
                  type: array
                  items:
                    type: object
                    properties:
                      url:
                        type: string
                        example: "http://img.com/1"
                      type:
                        type: string
                        example: "image"
                poll:
                  type: object
                  properties:
                    question:
                      type: string
                      example: "Zgadzacie się?"
                    options:
                      type: array
                      items:
                        type: string
                        example: ["Tak", "Nie"]
      responses:
        '201':
          description: Sukces
          content:
            application/json:
              example:
                postId: 1
                status: "success"
        '400':
          description: Złe żądanie (błąd walidacji)
        '500':
          description: Błąd zapisu, wywołano rollback PG
    get:
      summary: Pobiera listę postów
      description: Pozwala na wyszukiwanie w postach po hasztagu lub ID autora.
      parameters:
        - in: query
          name: authorId
          schema:
            type: integer
        - in: query
          name: hashtag
          schema:
            type: string
      responses:
        '200':
          description: Sukces
          content:
            application/json:
              example:
                - id: 1
                  authorId: 1
                  bodyPreview: "Wiadomość w systemie"
                  createdAt: "2026-05-08T12:00:00.000Z"
                  author:
                    username: "janedoe"
  /posts/{id}:
    delete:
      summary: Usuwa zadany post z bazy PG oraz odpala usuwanie kaskadowe z Mongo (lub Job)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usunięto
  /posts/{id}/reactions:
    post:
      summary: Dodaje nową reakcję
      description: Operacja idempotentna reagowania na post.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  example: 1
                type:
                  type: string
                  example: "LIKE"
      responses:
        '201':
          description: Reakcja zapisana
  /posts/{id}/comments:
    post:
      summary: Zwraca opcję dodania komentarza
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                authorId:
                  type: integer
                  example: 2
                parentId:
                  type: integer
                  nullable: true
                  example: null
                content:
                  type: string
                  example: "Komentarz pod postem!"
      responses:
        '201':
          description: Dodano pomyślnie
    get:
      summary: Pobiera drzewo komentarzy do postu
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sukces
          content:
            application/json:
              example:
                - id: 1
                  content: "Komentarz pod postem!"
                  replies: []
  /users:
    get:
      summary: Pobiera dane użytkowników testowych
      responses:
        '200':
          description: OK
  /users/{id}/follow:
    post:
      summary: Dodaje akcję obsewowania użytkownika (Follow)
      description: Z blokadą self-follow.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                followerId:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Obserwujesz.
    delete:
      summary: Cofa akcję obsewowania użytkownika (Unfollow)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                followerId:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Przestałeś obserwować.
  /feed/{userId}:
    get:
      summary: Pobiera prywatny zdenormalizowany feed użytkownika
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Pobrano feed (paginacja cursorem z MongoDB)
          content:
            application/json:
              example:
                data:
                  - postId: 1
                    score: 10
                    insertedAt: "2026-05-08T12:00:00.000Z"
                    richContent: 
                      attachments: []
                      poll: null
                nextCursor: "2026-05-08T12:00:00.000Z"
  /stats/user/{id}:
    get:
      summary: T1 - Statystyki użytkownika (pg native, parametryzowane $1)
      description: Endpoint korzystający z natywnego sterownika pg (Pool singleton, query $1).
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                user:
                  id: 1
                  username: "alice"
                  email: "a@x.pl"
                  created_at: "2026-05-08T12:00:00.000Z"
                stats:
                  postsCount: 3
                  followersCount: 5
                  followingCount: 7
                  reactionsReceived: 11
        '400':
          description: Niepoprawny ID
        '404':
          description: Użytkownik nie istnieje
  /stats/posts/top:
    get:
      summary: T1 - Top posty (pg native, $1 + $2)
      parameters:
        - in: query
          name: since
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: OK
  /tags:
    get:
      summary: T2 - Lista tagów z dynamicznym WHERE (Knex Query Builder)
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: minUsage
          schema:
            type: integer
        - in: query
          name: createdBefore
          schema:
            type: string
            format: date-time
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [usage_count, name]
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                count: 2
                tags:
                  - id: 1
                    name: "tech"
                    description: "Technologia, programowanie, narzędzia."
                    usageCount: 10
                    createdAt: "2026-05-14T12:00:00.000Z"
    post:
      summary: T2 - Utworzenie tagu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "tech"
                description:
                  type: string
                  example: "Technologia"
      responses:
        '201':
          description: Utworzono
  /tags/attach:
    post:
      summary: T2 - Przypisanie tagu do postu (transakcja Knex)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [postId, tagName]
              properties:
                postId:
                  type: integer
                  example: 1
                tagName:
                  type: string
                  example: "tech"
      responses:
        '201':
          description: Przypisano
  /tags/{name}/posts:
    get:
      summary: T2 - Posty z danym tagiem (JOIN w Knex)
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /notifications/{userId}:
    get:
      summary: T3 - Lista powiadomień użytkownika (Sequelize eager loading)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
        - in: query
          name: unread
          schema:
            type: boolean
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                count: 1
                notifications:
                  - id: 5
                    userId: 1
                    typeId: 2
                    message: "Ktoś polubił Twój post"
                    isRead: false
                    type:
                      id: 2
                      name: "reaction"
                      icon: "❤️"
                      priority: 4
  /notifications:
    post:
      summary: T3 - Utworzenie powiadomienia (walidatory Sequelize, hook beforeCreate)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, typeName, message]
              properties:
                userId:
                  type: integer
                  example: 1
                typeName:
                  type: string
                  enum: [follow, reaction, comment, mention]
                  example: "follow"
                message:
                  type: string
                  example: "Nowy obserwator"
                relatedPostId:
                  type: integer
                  nullable: true
      responses:
        '201':
          description: Utworzono
        '400':
          description: Walidacja Sequelize
        '404':
          description: Typ powiadomienia nie istnieje
  /notifications/{userId}/read-all:
    patch:
      summary: T3 - Oznacz wszystkie powiadomienia jako przeczytane (managed transaction)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                markedAsRead: 7
  /notifications/{id}:
    delete:
      summary: T3 - Usuń powiadomienie
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Usunięto
        '404':
          description: Nie znaleziono
  /analytics/top-authors-weekly:
    get:
      summary: Analiza MongoDB pobierająca najlepszych autorów z tego tygodnia (Aggregation Pipeline)
      responses:
        '200':
          description: OK
  /analytics/reaction-distribution:
    get:
      summary: Analiza rozkładu reakcji z systemu MongoDB i PostgreSQL (Złączone Pipeline)
      responses:
        '200':
          description: OK
  /analytics/system-logs:
    get:
      summary: System dzienników wspierany na natywnym sterowniku MongoDB (MongoClient)
      responses:
        '200':
          description: OK
    post:
      summary: Rejestrowanie logu systemowego z użyciem natywnego sterownika (T5)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                level:
                  type: string
                  example: "info"
                message:
                  type: string
                  example: "System reboot"
      responses:
        '201':
          description: Zapisano log
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object